Adjusts the BuildTreeScoped decompression cache factory to automatically close it's created caches when it is closed
Also refactors this and the ProjectScoped cache factory to share a common abstract base class which handles managing
the closings.  Makes the ProjectScoped factory thread-safe.
