+++ subprojects/core/src/main/java/org/gradle/api/internal/cache/BuildTreeScopedDecompressionCacheFactory.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2022 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.api.internal.cache;
+
+import org.gradle.cache.scopes.BuildTreeScopedCache;
+
+/**
+ * An implementation of {@link AbstractManagedDecompressionCacheFactory} that creates
+ * decompression caches.
+ */
+public final class BuildTreeScopedDecompressionCacheFactory extends AbstractManagedDecompressionCacheFactory {
+    private final BuildTreeScopedCache cacheFactory;
+
+    public BuildTreeScopedDecompressionCacheFactory(BuildTreeScopedCache cacheFactory) {
+        this.cacheFactory = cacheFactory;
+    }
+
+    @Override
+    protected BuildTreeScopedCache getScopedCache() {
+        return cacheFactory;
+    }
+}
